+++ src/main/java/com/verdantartifice/primalmagic/common/tiles/base/TileInventoryPM.java
@@ -33,6 +33,7 @@ public class TileInventoryPM extends TilePM implements WorldlyContainer {
     protected List<ContainerListener> listeners;    // Listeners to be informed when the inventory contents change
     protected final Set<Integer> syncedSlotIndices; // Which slots of the inventory should be synced to the client
     protected final int[] faceSlots;                // The slots of this tile's inventory visible from its sides
+    protected int ticksExisted = 0;                 // FIXME Remove when Forge onLoad bug is fixed
     
     public TileInventoryPM(BlockEntityType<?> type, BlockPos pos, BlockState state, int invSize) {
         super(type, pos, state);
@@ -238,6 +239,11 @@ public class TileInventoryPM extends TilePM implements WorldlyContainer {
     @Override
     public void onLoad() {
         super.onLoad();
+        // FIXME Uncomment when Forge onLoad bug is fixed
+        //this.doInventorySync();
+    }
+    
+    protected void doInventorySync() {
         if (!this.level.isClientSide) {
             // When first loaded, server-side tiles should immediately sync their contents to all nearby clients
             this.syncSlots(null);